Added a unit test for my directive and refactored some things.
