+++ templates/UserProfile.html
@@ -14,7 +14,7 @@
         <button id="Edit" ng-click="enableEditor()">Edit</button>
 
         <!--Adding in the userInfo with a directive-->
-        <user-info ng-show="userInfo" id="Text" user="user"></user-info>
+        <user-info ng-show="userInfo" id="Text"></user-info>
 
     </div>
