--- templates/directives/userInfo.html
@@ -1,6 +1,0 @@
-<p style="font-size: 18px;">
-	<span>First Name: {{user.firstName}}</span>
-	<span>Last Name: {{user.lastName}}</span>
-	<span>Phone: {{user.phone}}</span>
-	<span>Email: {{user.email}}</span>
-</p>
