+++ test/e2e/conf.js
@@ -1,4 +1,4 @@
 exports.config = {
   seleniumAddress: 'http://localhost:4444/wd/hub',
-  specs: ['Spec.js']
+  specs: ['e2eSpec.js']
 };
