+++ test/unit/directivesSpec.js
@@ -1,0 +1,25 @@
+describe('Unit testing userInfo', function() {
+  var $compile,
+      $rootScope;
+
+  // Load the myApp module, which contains the directive
+  beforeEach(module('sampleApp'));
+
+  // Store references to $rootScope and $compile
+  // so they are available to all tests in this describe block
+  beforeEach(inject(function(_$compile_, _$rootScope_){
+    // The injector unwraps the underscores (_) from around the parameter names when matching
+    $compile = _$compile_;
+    $rootScope = _$rootScope_;
+  }));
+
+  it('Replaces the element with the appropriate content', function() {
+    // Compile a piece of HTML containing the directive
+    var element = $compile("<user-info></user-info>")($rootScope);
+    // fire all the watches, so the scope expressions will be evaluated
+    $rootScope.$digest();
+    // Check that the compiled element contains the templated content
+    expect(element.html()).toContain("First Name:", "Last Name:", "Phone:", "Email:");
+    expect(element.html()).toContain("ng-binding");
+  });
+});
